requirements.txt
django==5.0.3
djangorestframework==3.15.0
drf-yasg==1.21.7
python-dotenv==1.0.1
pytest==8.1.1
djangorestframework-simplejwt==5.3.1
